DO NOT DOWNLOAD. DOWNLOAD LATEST STABLE FROM "DOWNLOADS" TAB - Entity framework. Further changes. - Removed bloated data access

+++ Libraries/Nop.BusinessLogic/Orders/OrderAverageReportLine.cs
@@ -43,7 +43,7 @@ namespace NopSolutions.NopCommerce.BusinessLogic.Orders
         /// <summary>
         /// Gets or sets the count
         /// </summary>
-        public decimal CountOrders { get; set; }
+        public int CountOrders { get; set; }
 
         #endregion
     }
