+++ Libraries/Nop.BusinessLogic/Orders/OrderAverageReportLineSummary.cs
@@ -47,7 +47,7 @@ namespace NopSolutions.NopCommerce.BusinessLogic.Orders
         /// <summary>
         /// Gets or sets the today count
         /// </summary>
-        public decimal CountTodayOrders { get; set; }
+        public int CountTodayOrders { get; set; }
 
         /// <summary>
         /// Gets or sets the sum this week total
@@ -57,7 +57,7 @@ namespace NopSolutions.NopCommerce.BusinessLogic.Orders
         /// <summary>
         /// Gets or sets the this week count
         /// </summary>
-        public decimal CountThisWeekOrders { get; set; }
+        public int CountThisWeekOrders { get; set; }
 
         /// <summary>
         /// Gets or sets the sum this month total
@@ -67,7 +67,7 @@ namespace NopSolutions.NopCommerce.BusinessLogic.Orders
         /// <summary>
         /// Gets or sets the this month count
         /// </summary>
-        public decimal CountThisMonthOrders { get; set; }
+        public int CountThisMonthOrders { get; set; }
 
         /// <summary>
         /// Gets or sets the sum this year total
@@ -77,7 +77,7 @@ namespace NopSolutions.NopCommerce.BusinessLogic.Orders
         /// <summary>
         /// Gets or sets the this year count
         /// </summary>
-        public decimal CountThisYearOrders { get; set; }
+        public int CountThisYearOrders { get; set; }
 
         /// <summary>
         /// Gets or sets the sum all time total
@@ -87,7 +87,7 @@ namespace NopSolutions.NopCommerce.BusinessLogic.Orders
         /// <summary>
         /// Gets or sets the all time count
         /// </summary>
-        public decimal CountAllTimeOrders { get; set; }
+        public int CountAllTimeOrders { get; set; }
 
         #endregion
     }
